move check-email-already-exists into domain-service
