+++ src/config/participant-router.ts
@@ -5,7 +5,7 @@ import { createContext } from "../infra/shared/context"
 import { ParticipantRepository } from "../infra/participant/participant-repository"
 import { PairRepository } from "../infra/pair/pair-repository"
 import { TeamRepository } from "../infra/team/team-repository"
-import { CheckEmailAlreadyExists } from "../domain/participant/check-email-already-exists"
+import { CheckEmailAlreadyExists } from "../domain/participant/domain-service/check-email-already-exists"
 import { ParticipantFactory } from "../domain/participant/participant-factory"
 import { PairFactory } from "../domain/pair/pair-factory"
 import { GeneratePairName } from "../domain/pair/domain-service/generate-pair-name"
