+++ src/domain/participant/participant.ts
@@ -2,7 +2,7 @@ import { Entity } from "../shared/entity"
 import { ParticipantId } from "./participant-id"
 import { ParticipantName } from "./participant-name"
 import { Email } from "./email"
-import { CheckEmailAlreadyExists } from "./check-email-already-exists"
+import { CheckEmailAlreadyExists } from "./domain-service/check-email-already-exists"
 
 export interface ParticipantProps {
   name: ParticipantName
